requirements.txt
Django==3.2
whitenoise==5.2.0
gunicorn==20.1.0
psycopg2==2.8.6
djangorestframework==3.12.4
django-rest-auth[with_social]==0.9.5
django-heroku==0.3.1
geoip2==4.1.0
